Cancelable rechecks test
Summary:
Introduces a basic test for cancelable rechecks. The general idea is that we have a file which takes a long time to check (thanks to `$Flow$DebugSleep`). During the recheck, I'll remove it. The initial recheck will be canceled and the new one will run quickly.

This test also demonstrates that initialization cannot be canceled.

Reviewed By: pakoito

Differential Revision: D8789885

fbshipit-source-id: 33ecb69031aa0178f5fdc4ea816e8502da8be670