always create a top-level var scope in Scope_builder
Summary:
ignore_toplevel is used to bypass creating a scope containing top-level variables, so that they get treated like globals, which is what you want when parsing a script (but not a module, where top-level stuff is module-private).

however, we assume that lexical scopes always have a parent, and not creating the top-level scope violates that assumption. creating such a scope with no bindings has the same effect as not creating it, but maintains the invariant.

Reviewed By: pakoito

Differential Revision: D8816018

fbshipit-source-id: 174c39ad40a8967022f8e92c17b9e3e89b8b6bca