Add second type parameter to AST for nodes that can be typed
Summary:
Preparing the AST to support sometimes being type-annotated. Previously the AST was parameterized on a single annotation type which all nodes were annotated with. However this doesn't make sense for type annotations, as not every node in the AST is meaningfully associated with a type.

To fix this, I've instead parameterized the AST on two type variables: one for annotations on "untypeable" nodes and one for annotations on "typeable" nodes.

The parser used to return a `Loc.t Ast.program`, and now it returns a `(Loc.t, Loc.t) Ast.program`. The goal is for constraint generation to then transform this into a `(Loc.t, Loc.t * Type.t) Ast.program`.

`parser/ast.ml` is the most important file change in this diff, and it shows precisely which sorts of nodes in a `('M, 'T) Ast.program` are annotated with `'M` and which are annotated with `'T`. Currently every node continues to be annotated with `'M` except for `Ast.Expression.t` and `Ast.Type.t`, which are annotated with `'T`. If in the future we decide that there are other nodes that it also makes sense to type-annotate, we can update `ast.ml` to annotate those other nodes with `'T` instead of `'M`.

All the other file changes in this diff are just to update type annotations to be consistent with the new arities of the AST type constructors. There are only type changes in this diff; no expressions have changed, so there should be no performance impact.

Reviewed By: samwgoldman

Differential Revision: D8945785

fbshipit-source-id: aa994c444c6a3d9648fe94e5a74122df71c361c5