Polymorphic ast mapper
Summary:
Now that we're starting to use the AST with annotations other than `Loc.t`, it'll be useful to have a mapper that can change the annotation type. Like with the original mapper, this mapper can be customized by inheriting from it. Unlike the original mapper, this mapper is instantiated with a function which transforms the annotation type.

I copied and edited the source code from `flow_ast_mapper`. There were some unimplemented TODOs that would just return the original AST in certain cases; since this wouldn't typecheck when transforming the annotation type, I implemented those TODOs. These implementations can't be copied back to `flow_ast_mapper` verbatim, because we'd also have to add the physical equality optimizations.

Reviewed By: samwgoldman

Differential Revision: D8813926

fbshipit-source-id: 7c22628babfe8811cc565d61a175c2047c974730