--flowconfig-name to specify a name for a flowconfig
Summary:
As the title-- you can now specify the name of the flowconfig.

Unfortunately, there was no good place to add the flag so that it would be added to all the commands that would need it, so I created such a set of flags and named it `Base_flags`.

I added `flowconfig_name` to `Options` to minimize the amount of times I would need to explicitly pass it as an argument.

In order to handle running multiple servers at a time, we add the flowconfig's name to any server generated files. If it's just .flowconfig, we do not do anything. This is to make sure we don't break anything depending on these filenames.

Reviewed By: gabelevi

Differential Revision: D8899890

fbshipit-source-id: f8ba4e9393034ffe41120ca0b2cb03b39a026180