Unit testing typed AST preservation of AST structure
Summary:
Added a system for unit testing the typed ASTs produced by constraint generation. The testing framework takes a source file and runs it through constraint generation to get a typed AST; then pretty-prints the original AST to a file and the typed AST to another file, and diffs them. If a difference is detected the test fails and the diff output is shown.

Not sure if I'm following best practices wrt file paths; for example the `diff_dif` used in lines 120-135

Reviewed By: panagosg7

Differential Revision: D8897479

fbshipit-source-id: e9e2e126edfeceb0ac698907e8a669adde804dba