Fix typed AST translation of DeclareFunctions without predicates
Summary:
Previously `Statement.statement` would only properly translate a DeclareFunction if it had a predicate, as this was the only case in which the type annotation would get translated in `statement` rather than `statement_decl`.

For almost all statements we've been able to only do the translation in the `statement` function (the second pass) and never need to return information from the `statement_decl` function (the first pass). I didn't want to change the code to also return information `statement_decl`, as that would make things a lot more complicated just to deal with one case.

Instead, I updated environments to optionally hold a type annotation. Since the first pass through a DeclareFunction populates the environment with the function's type anyway, I only had to update this to also store the function's type annotation's typed AST. Then this typed AST is retrieved from the environment during the second pass.

Side note: this allows us to use the typed AST unit testing (see previous diff) for the first time!

Reviewed By: panagosg7

Differential Revision: D8908069

fbshipit-source-id: 3eb8a5ad7e13b760aed97c433c5bca55341b5b09