Typed AST: annotate Ast Type and Expression nodes with Loc.t * Type.t
Summary:
Puts the "typed" in "typed AST". Every Type and Expression node in the AST returned by the typechecker is now annotated with `Loc.t * Type.t`. All other nodes are still annotated with `unit`. An upcoming diff will change those to be annotated with `Loc.t`.

An issue that came up when writing this diff is that there are some nodes that types aren't constructed for (for example `requireLazy`, `module`, etc). I've marked these as TODOs for the time being, and they can be filled in a future diff once I learn how best to construct these types.

This diff changes a few tests because it makes typechecking happen in a few places that it didn't happen before. In particular, it changes the `requireLazy` test because it now traverses non-string elements of the array passed to `requireLazy`. It changes the `dump-types` test because it now traverses the string literals passed to `require`.

Both of these changes seem fine to me (and unlikely to majorly break anything as far as I can tell), so ~~I'm inclined to make these tests pass by updating~~ I updated the `.exp` files. Another option would have been to just make the typechecker not traverse these new things, and return `Any` in these cases.

Reviewed By: panagosg7

Differential Revision: D9091177

fbshipit-source-id: 79d9511b5055e678f2115614c2e5765e7fc2cac2