Typed AST: annotate non-typed nodes with Loc.t
Summary:
Completes the Typed AST project!

Now every node in the AST returned from the typechecker will be location-annotated, which makes it so that the typed AST isn't missing any information that the original AST had. This will, for example, make it easier to search the typed AST for a specific position, which will be useful for porting type-at-pos over to the typed AST.

Reviewed By: panagosg7

Differential Revision: D9126402

fbshipit-source-id: 296ad30b0b26add9b3230857bffba592255d1955