Watchman init can specify starting clock
Summary:
See D8431351 for full story of where we will end up and how
the "dirty files" calculation will be split up during an
Informant-directed restart.

As a requirement, Watchman.init will need to support starting a
subscription from a given clock. This is supported by Watchman ( See docs at
https://facebook.github.io/watchman/docs/cmd/subscribe.html ).

Quoting the docs: "The subscribe command object allows the client to specify a since parameter;
if present in the command, the initial set of subscription results
will only include files that changed since the specified clockspec,"

Reviewed By: dabek

Differential Revision: D8554723

fbshipit-source-id: 47a8b263ebb4d7eb03026ab9944d9e4710ac155b