Fix flaky TAST diffing test
Summary:
The Typed AST diffing test I put up was flaky because running multiple invocations of the same test from the same flow root at the same time would cause them to overwrite each other's files. This diff changes it to append a random number to each test run's directory, which decreases the likelihood of the test runs stepping on each other.

Technically it's still possible (but very unlikely) for the different test runs to have the same random number, so this doesn't 100% fix the test's flakiness. Is there a better way to get a unique number for each run?

Reviewed By: panagosg7

Differential Revision: D9197973

fbshipit-source-id: 43a1ea5278935cd516ef3f01250d1e9eeea1cdef