Profiling_js refactor/rewrite
Summary:
== Problem ==

`Profiling_js` was written to profile stuff like

* Profile init
  * Profile parsing
  * Profiling merging
  * etc

So we had one big slow thing and we wanted to profile each individual step.

But now we find outselves with much more complicated code paths. Stuff like

* Profile type-at-pos command
  * typecheck_contents -> ensure_checked_dependencies
      * realize unchecked dependencies are unparsed and hashes don't match & start a recheck
        * Profile parsing
        * Profiling merging

How do we deal with timers within timers? And what do we do when we need to call a function that is so slow that it returns its own profiling object?

== Solution ==

Refactor `Profiling_js` to support hierarchical profiling. That is:

1. Support running timers inside of other timers
2. Support merging a finished profiling object (like from a recheck) into the currently running profiling object (like from handling a command)

Reviewed By: samwgoldman

Differential Revision: D8980471

fbshipit-source-id: 8ac5b023eb2419a434865c0778cb00ae0d403ee5