Simplify the calling of ensure_checked_dependencies
Summary:
I personally found `ensure_checked_dependencies` a little confusing. Rather than being called directly inside of `types_js.ml`, it was passed as a callback to `merge_service.ml` and invoked later. I imagine this was done to make sure it was always called.

However, as the APIs evolved, it became the user's prerogative whether or not `ensure_checked_dependencies` was called. So the guarantee weakened.

I feel like calling it directly makes it easier to understand when it runs.

Reviewed By: nmote

Differential Revision: D9073383

fbshipit-source-id: 33183486d0db220e109cf70fc7497165a7f8cf28