[Easy] Enforce singleton mutators
Summary:
A transaction keeps track of 0 or more mutators, each of which is committed or rolled back at the end of the transaction. Mutators often need to keep a copy of the old data, so that it can be restored on a rollback. Sometimes this is stored in a local variable. But sometimes this is stored globally, like oldified shared memory data.

That means some mutators MUST be singletons. For these mutators, if two of the same mutator are active at the same time, they could overwrite the stored old data.

This diff goes through the mutators, figures out which ones need to be singletons, and enforces it.

Reviewed By: pakoito

Differential Revision: D9114740

fbshipit-source-id: 3baf2d47f7303565d3c38a1ed8b42865cd88ad25