Cancelable workloads (i.e. commands)
Summary:
So we have cancelable rechecks. When we learn that a file has changed in the middle of a recheck, we cancel that recheck and spawn a new recheck which includes the new changes.

This diff introduces cancelable workloads. A workload is something like handling a command. If we learn that a file has changed while we're handling a command, we stop handling that command, perform a recheck, and then re-run the workload from scratch.

This helps enable `ensure_parsed`. For example, consider this situation:

*  We're handling a `type-at-pos` command
* Due to lazy mode, a dependency `foo.js` is not yet checked
* Due to saved state `foo.js`'s AST is not in shared memory
* When we go to read `foo.js`, we notice that it has changed but no one has told us yet

With cancelable workloads, we can then cancel the `type-at-pos` command, perform a recheck for `foo.js`, and then re-run the `type-at-pos` command.

Reviewed By: nmote

Differential Revision: D9114738

fbshipit-source-id: 6bc3c71630bacc8482fa2dc86f9c9e899cc4a4e4