ensure_parsed
Summary:
Initializing a server using saved state will populate many of the heaps, but it will skip parsing and won't populate the `ASTHeap` and the `DocblockHeap`. If we intend to merge a file after loading with a saved state, we need to actually parse that file.

So immediately before we merge, we call `ensure_parsed`, which checks the `ASTHeap` for each file we're about to merge and parses those which are missing.

Reviewed By: samwgoldman

Differential Revision: D8488316

fbshipit-source-id: 302f1baafc8d34a4848f4cac2bf427e11a233281