[Easy] Add file hashes to the saved state
Summary: We need to be able to detect if a file has changed on disk without the file-watcher telling us. This can happen if the file watcher is turned off or if the file watcher notification is delayed. Saved state loading needs this, since we need to know if the saved state info (which is sparse - not all shared memory is serialized in the saved state) is out of date so that we can trigger a recheck.

Reviewed By: jbrown215

Differential Revision: D9114804

fbshipit-source-id: a9fe84a602bc9ce0f89e85032ecb29378caea40b