Test saved state with `runtests.sh -s`
Summary:
Flow should behave the same way, whether or not it used saved state to initialize. The tests in `tests/` do a pretty good job of testing a whole bunch of edge cases. So it makes sense to use this test suite to make sure saved state works.

`./runtests.sh -s` will run a little slower. The basic idea is that we'll replace `flow start` and `flow check` with `flow start --shared-state-load-script foo.sh` and `flow check --shared-state-load-script foo.sh`. And before we can run these commands we need to

* Start a lazy flow server
* Create a saved state file
* Stop the server
* Create a saved state load script

Reviewed By: mroch

Differential Revision: D8504425

fbshipit-source-id: 7bb1d16e6728f5ee18d3bd843248f00dc835c5d0