Send server logging context to lsp
Summary:
One of the downsides to moving the persistent command logging
to the lsp command process is that the logging context is now wrong (or
at least different). You no longer can tell stuff like whether the
server was running in lazy mode or what file watcher was being used.

This fixes that issue by sending the server logging context to the lsp
command and using that when logging the success and failure events
instead of the lsp command's context (which is logged as the client
context)

Reviewed By: ljw1004

Differential Revision: D9129401

fbshipit-source-id: 7a3ef9c956cfad1ede6ff82f8241864f5b673921