Use marked instead of ISet in the assert_ground_visitor
Summary:
In our assert_ground_visitor, we maintain a `seen` set in order to prevent us from revisiting an OpenT we've already seen. This does two things:
1. Breaks cycles
2. Caches work that would otherwise be repeated

However, our `seen` set did not take into account the polarity when these `OpenT`s were seen. This means that we might see an `OpenT` in a covariant position and add it to our seen set , but never complain later even if we end up seeing it in a contravariant position.

Here's an example that illustrates this bug:
https://flow.org/try/#0CYUwxgNghgTiAEAzArgOzAFwJYHtXwxAGcMAeAFQD4AKASgC55rrz4BeS+c29z7gbgBQgkAA8ADjhgZ4oRFGQQZhEnX5A

The first time we encounter `T`, it is in a covariant position. However, the second time `T` appears it is in a contravariant position. We ought to complain about an error here. If you change the first `T` to `any`, we do indeed error!

Instead of using an ISet, we should use the Marked set (originally made for gc_js.ml). This set is sensitive to the polarities, which are specified when inserting. Now, if we ask if we've seen an given OpenT, we also ask if we've seen it under this specific polarity.

Reviewed By: samwgoldman

Differential Revision: D9241834

fbshipit-source-id: a1894d806e95a46ced876d0f7a6027a364f9f83d