Create an API for transforms to create new variables
Summary:
A lot of transforms need to create new variables amd these variables need to be declared in the scope they are used. This involves a work of tracking the scope while traversing along the AST.

To avoid re-implementing this process while implementing different transforms, we have created the `Variable_controller` class with 3 public methods:
- `create_new_var`: create a new variable and declare it at the nearest block
- `make_block_visit_function` create a function to overwrite the block traversing function of `Flow_ast_mapper`
- `make_program_visit_function` same as above, but for traversing the whole program.
Every transform which needs to create new variables, just creates one instance of this controller, and overwrite the block and program visiting function.

Reviewed By: pakoito

Differential Revision: D9133383

fbshipit-source-id: 9844d6e8968fafea228f016588376748e22012f6