[PR] Fix docs for `$Shape`
Summary:
Docs for `$Shape` were introduced in #6694, but that PR had four major
errors:

  - the section was not linked from the table of contents;
  - the type `$Shape<T>` was misspelled as `Shape<T>` in the header;
  - the code block did not have `// flow`, so lacked error annotations;
  - the code block was not terminated, so the entirety of the rest of
    the page was marked as code, [like this][1]—clearly no one previewed
    the page before deploying!

[1]: https://archive.fo/CqH7D#87%

This patch fixes those errors. While in the area, we clean up the
comments reading "Error" and "OK" to (a) eliminate the horizontal
scrollbar and (b) match the style elsewhere on the page.

Closes #6692.
Pull Request resolved: https://github.com/facebook/flow/pull/6708

Reviewed By: panagosg7

Differential Revision: D9249420

Pulled By: jbrown215

fbshipit-source-id: a54de957c5ac744ce7c483c4be0cbfd182dfee02