Remove try-with that was catching Lwt.Canceled
Summary:
On the master thread, there was a `try...with` around the merge code. This is not to be confused with [the one in merge_service.ml](https://github.com/facebook/flow/blob/27282dd12d0709f3d71c15eba3d5ff063567136d/src/services/inference/merge_service.ml#L268) which is on the worker thread and turns exceptions into Flow errors. I'm honestly not sure what good this `try...with` is doing any more, so I'm going to remove it. It's very old.

It is doing some harm, though.

* In v0.78.0 we forgot to let `Lwt.Canceled` through, which broke cancelable rechecks
* It catches exceptions before they hit the transaction barrier, which means the transaction is committed and not rolled back. This easily leads to bad state
* It's just printing out the exception. I can't think of that ever helping.

Reviewed By: avikchaudhuri

Differential Revision: D9276209

fbshipit-source-id: 6c9cb9a1c8c49920f92fa4674a95be52be4284b6