[Normalizer] Refactor and make API clearer
Summary:
This diff does not add new functionality to the normalizer, but improves
in the following ways:

* **Normalizer Environment** (`ty_normalizer_env.mli`): Adds a separate file that hold
the environment. Its fields are:
  * Global part (`genv`): remains invariant during normalization and holds
    * File of query
    * Context (fully resolved): this is the context of the leader of a merged component
    * Current module imports

  * Normalization `options`

  * Available imports

  * Rest: parts that may differ in either of the axes
    * location in file --> type parameters in scope
    * depth of normalization --> type alias resolution info

* **Normalizer API** (`ty_normalizer.mli`): all callbacks require `options` and `genv`.
This will be useful soon whenmore fields will be required in `genv`, that will have to be explicitly
determined. E.g. see D9260484 where the type-table & `imported_ts` may not correspond
to the merged context.

Reviewed By: gabelevi

Differential Revision: D9260485

fbshipit-source-id: d71ad12db5ffb897e3251f9bef49e748b8637f57