[Normalizer] Make File_sig and type-tables available to the normalizer
Summary:
This diff sets up the normalizer inputs to use `File_sig` information to compute imported
names (instead of the current `Context.imported_ts`). This is currently one of the last
dependencies of `imported_ts`, which will bring us closer to deprecating it. File_sig ought
to provide all of its functionality.

The normalizer will use this information in D9260483.

Reviewed By: gabelevi

Differential Revision: D9260484

fbshipit-source-id: 58085e46ee61f3656a0cbb7d159e89e071374b19