Speed up ensure_parsed
Summary:
`ensure_parsed` makes sure that every file we're about to merge has been parsed into `ASTHeap`. For a full check, this means EVERY file. And we were doing this one by one on the master thread, which was taking around 45s. If we do this on worker threads, it will instead take 2s.

I did explore some other solutions:

* Keeping track in `Parser_heaps` which files are known but haven't been parsed yet due to saved state. This would add some more global state to the master process, so we'd need to be able to roll it back via a transaction.
* Keeping track in env which files haven't been parsed yet. This avoids global state but we have to make sure to update the env whenever a file is successfully parsed.

Reviewed By: panagosg7

Differential Revision: D9347529

fbshipit-source-id: 03dd793a78c0db6abda4e38b9997341a5285d911