Ensure that field locations are always present
Summary:
I want to use these locations to create tvars, which means they must be
present. While the class sig stores fields with the possibility of a missing
location, the locations are always present at this point in the code.

One exception is the static name field which all classes have implicitly. We
don't need to create a tvar for the name field for our purposes.

Follow-up diffs will make use of these locations.

Reviewed By: panagosg7

Differential Revision: D9310344

fbshipit-source-id: 6658f174a0d491f9c4d6632c9db5ac239552ec9e