[type-visitor] Pass down pole when visiting a TypeDestructorTriggerT
Summary:
`TypeDestructorTriggerT`s carry around a tout into which the result of the type destructor
evaluation eventually flows. Before this change, we were conservatively walking the tout with a
neutral polarity. However, this is unnecessarily strict. Whenever a TypeDestructorTriggerT appears as
an upper bound the tout remains in the upper bound of the newly emitted constraint.
Similarly, whenever a TypeDestructorTriggerT appears as a lower bound the tout remains in the lower
bound of the new constraint as well.

Reviewed By: samwgoldman

Differential Revision: D9360122

fbshipit-source-id: f901ca00174f66628dbd7eecbc252c4bf1e78e4f