Introduce new AST node for class extends
Summary:
Before this change, a class declaration had separate `super` and `super_targs`
fields. This change wraps those in a new AST node, which has two main benefits:

1. The relationship that `super = None` implies `super_targs = None` is now
enforced by the types. I was able to remove one explicit assert.

2. There is now a location for the entire extends-expression-with-type-
arguments, which is a meaningful location for the type application. This
removes a Loc.btwn call in statement.ml

Reviewed By: gabelevi

Differential Revision: D9350943

fbshipit-source-id: c6ee99e69eadf6121ca2411c9df74422c1414a8e