signature builder deps
Summary:
While building an environment, we don't "evaluate" any bindings. We do that later, and the way it is
done is by returning a set of dependencies that a binding refers to. A dependency might demand
that a name is a type or a value, or that it is an import of a particular kind from
a specified source, or that it is a global type or value.

Later, we can recursively resolve named dependencies by looking up the names in the environment and
evaluating the corresponding bindings.

Reviewed By: pakoito

Differential Revision: D9310858

fbshipit-source-id: 3f4b489797f0ebe6aeca35c7072906cd707de6ea