Slimmer BoundTs
Summary:
Before this diff, a BoundT stored the entire typeparam record, which includes
the bound and default type.

The BoundT only needs to store a reason and a name, because we will look up the
type params from the map during substitution.

I left in the polarity, but that is technically not necessary either. The
polarity is accessed during polarity checks, which do not currently have access
to type parameters in scope.

Leaving the bound/default types should be harmless, except that the type
visitor will actually visit these types unnecessarily, and those types will
also appear unnecessarily in signature contexts.

Reviewed By: avikchaudhuri

Differential Revision: D9445076

fbshipit-source-id: 8d11358ed3178f0ba822de3f23dfefc6fa5536d1