Replace ad-hoc no-this-before-super logic with TDZ-based logic
Summary:
It is illegal to reference `this` and `super` before calling `super()` in a
derived constructor, because the superclass is responsible for allocating these
objects.

Before this change, we enforced this by making these bindings optional at the
start of a derived constructor, then refining them to non-optional variants
after the super call.

This logic is problematic for a couple reasons.

One reason is performance. Every time we see a super() call, we emit a
constraint `T <: ?T`, which is trivially true but potentially expensive,
depending on the particular `T`.

The other reason is simple correctness. Modeling `this` as an optional type
means you could write code like this:

```
class A {}
class B extends A {
  constructor() {
    this && this.foo;
  }
}
```

The above passes the type checker (because we refine `?T` to `T`) but will fail
at runtime, because even referencing `this` is illegal.

The existing TDZ infra is perfectly capable of handling this case as well. In a
derived constructor, we create entries for `this` and `super`, but they start
out in the `Undeclared` state. When we see a constructor call, we flip the
state to `Declared`, then to `Initialized`.

This caused issues with `mk_arrow`, which re-uses the existing `this` and
`super` entries, but did so by reading those entries from the environment,
which caused TDZ even attempting to create an arrow function. I worked around
this by exposing an API to get the raw Entry.t from the environment instead.

Reviewed By: gabelevi

Differential Revision: D9369849

fbshipit-source-id: 876ed2992ce3ec1bada51aa858e98a112b115ea1