Eliminate unnecessary "for value" case in Flow_js.mk_instance
Summary:
This codepath exists to turn a `ClassT i` into `i` and is only ever called when
evaluating the result of a `ThisTypeAppT`.

The last step of this evaluation uses `ThisSpecializeT` which turns a
`ThisClassT i` into a `ClassT i`. Notice that this `ClassT` is then immediately
unwrapped.

So, let's just not wrap it up in the first place! This means we no longer need
to call `mk_instance`. Also, this scheme repositions the instance to the
position of the typeapp, which makes it unnecessary to unwrap the RStatics
reason description in check_super.

Reviewed By: panagosg7

Differential Revision: D9357525

fbshipit-source-id: fb64c1ac1849891d4835707275db5cc30e13f940