Fix opaque type unwrapping bug in React createElement code
Summary:
When we evaluate a createElement call, we combine the provided config object
with any default props the component provides. This combining uses a special
CondT use type to check if the incoming config value is void.

This use type used to cause opaque types to unwrap, which would cause bogus
errors.

Reviewed By: jbrown215

Differential Revision: D9390144

fbshipit-source-id: 00ce8414fb7ab14e845b48bdf43ed2234d6a1a2a