Use RClass description for ClassT/ThisClassT instead of RStatics
Summary:
ClassT types often appear in type positions, like in a type annotation. Errors
involving misuse will confusingly say "statics of C." This changes the reasons
in those positions to say "class C" instead.

It's important to say "class C" instead of just "C" because just "C" is the
description of _instances_ of C. If you try to use an instance when the class
type is expected, it would be confusing to say "C is incompatible with C."

When a class is used as an object, we DO want to say "statics of C," but that
is already working, because the instance contains an object type describing
those statics which already uses the RStatics description.

Reviewed By: jbrown215

Differential Revision: D9384102

fbshipit-source-id: 39c0014ad1fee38358e69ad0ff16a8c6eb34f5be