Make typeapp repositioning consistent
Summary:
Any time `C<X>` appears, we should consistently position the resulting type at
the location of the typeapp itself.

This is done fairly inconsistently today, involving ad-hoc repositioning of `C`
in some places, or repositioning later, or not repositioning at all.

This diff consolidates all repositioning for `TypeAppT` and `ThisTypeAppT`
types.

In the `TypeAppT` case, the repositioning happens in `mk_instance` which
creates an `AnnotT` type with the typeapp reason. The `AnnotT` will perform
repositioning later.

In the `ThisTypeAppT` case, the repositioning happens as part of
`ThisSpecializeT`, which uses `continue_repos` which will emit a `ReposLowerT`
or `ReposUseT` constraint as appropriate.

In other words, repositioning is lazy, occuring when the type application is
actually performed. Until then, the unapplied type maintains its original
position.

Reviewed By: jbrown215

Differential Revision: D9373917

fbshipit-source-id: 4df8df9af4795d5cee1460e01eaf15514398d838