Add Flow_js.flow_p to check property compatibility
Summary:
Internally, we have rec_flow_p, which is used for object compatibility and
others. This function requires a trace argument, which callers from outside
Flow_js do not have.

This refactor adds a `flow_opt_p` function which does not require the trace
object, and calls that from `rec_flow_p` which requires the trace. I also added
the public function.

In the next diff I will use this API to enforce that own properties on class
instances are compatible with proto properties declared in the same class.

Reviewed By: gabelevi

Differential Revision: D9438039

fbshipit-source-id: ee1d4be87aee45553189ccee31c16bf22529bc87