Enforce that own properties are compatible with proto properties in a class
Summary:
A class declaration simultaneously declares three separate shapes: the class
statics, the prototype object, and the owned properties of instances. The last
is done via the class properties proposal.

We already assert that own properties of a subclass are subtype compatible with
proto properties of a superclass. The same principle applies to own and proto
properties declared within the same class.

Reviewed By: panagosg7

Differential Revision: D9439853

fbshipit-source-id: bdaa740458e882a9bc84a0ece99982a971d9b2f2