Fix lookup for instance/object subtyping with optional properties
Summary:
Before this change, the added test caused a false positive type error:

```
Cannot cast new B() to object type because undefined [1] is incompatible with number [2] in property x.

 [1] 1│ declare class A { x?: number }
     2│ declare class B extends A {}
 [2] 3│ (new B: { x?: number }); // OK
     4│
```

This happened because the optional type in the upper bound was unwrapped before
looking up the property on the lower bound.

Also, one branch unnecessarily emitted a recursive constraint involving `l`,
when we already know the property does not exist in `l`. We should always
perform a lookup on the superclass.

Reviewed By: jbrown215

Differential Revision: D9481858

fbshipit-source-id: f94a400f75f8178ea059225fb5a73b980b094f90