[Easy] 2 saved state fetchers: dummy and local
Summary:
A saved state fetcher is a module which provides a single `fetch` function. When called, it returns two things:

1. The path to the saved state file
2. The set of files which have changed since the saved state was generated.

This diff introduces 2 saved state fetchers

This is a saved state fetcher which always fails to fetch the saved state

This is a saved state fetcher intended to be used for tests. It assumes that, next to the `.flowconfig`, there is also a `.flow.saved_state` and a `.flow.saved_state_file_changes`. The former is the saved state file and the latter is a newline-separated list of files which have changed.

Reviewed By: samwgoldman

Differential Revision: D9323271

fbshipit-source-id: 80553c7f2bec3fea1c7717d970bf5a83f0abbe08