Pass locs in Flow_ast_mapper
Summary:
So here's the deal, many of our utils use `loc`'s to make associations with AST nodes (the scope API is an example) but the `Flow_ast_mapper` makes it really hard to get access to the `loc` of a particular node. In a lot of cases we end up overriding `expression` or `statement` and matching just to get access to this, which means we lose a lot of value from the mapper. Additionally we already expose `loc`s for some nodes but not others so the methods expose an inconsistent API.

This diff fixes the above problems by adding the `loc` to all the mapper methods that have a loc for that node (so all of `expression` and `statement`, then some others like `type`).

Reviewed By: panagosg7

Differential Revision: D9476937

fbshipit-source-id: 0739f05dab4f0028c645ba9fd7715a449a00c17b